LayoutTests/platform/glib/imported/w3c/web-platform-tests/geolocation-API/enabled-by-permissions-policy.https.sub-expected.txt

CONSOLE MESSAGE: Permission policy 'Geolocation' check failed for document with origin 'https://not-web-platform.test:9443'.


PASS Permissions-Policy header geolocation=* allows the top-level document.
PASS Permissions-Policy header geolocation=* allows same-origin iframes.
FAIL Permissions-Policy header geolocation=* allows cross-origin iframes. assert_true: geolocation expected true got false
